+++ WegliKit/Sources/RegulatoryOfficeMapper/Mock.swift
@@ -1,0 +1,7 @@
+import ComposableArchitecture
+
+public extension RegulatoryOfficeMapper {
+  static let noop = Self(
+    mapAddressToDistrict: { _ in .none }
+  )
+}
